Move yaml helping functions to __init__.py
